+++ pmd-visualforce/src/main/java/net/sourceforge/pmd/lang/vf/DataType.java
@@ -14,8 +14,6 @@ import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import apex.jorje.semantic.symbol.type.BasicType;
-
 /**
  * Represents all data types that can be referenced from a Visualforce page. This enum consolidates the data types
  * available to CustomFields and Apex. It uses the naming convention of CustomFields.
@@ -104,42 +102,6 @@ public enum DataType {
         }
 
         return dataType;
-    }
-
-    /**
-     * Map to correct instance, returns {@code Unknown} if the value can't be mapped.
-     *
-     * @deprecated Use {@link #fromTypeName(String)} instead.
-     */
-    @Deprecated
-    public static DataType fromBasicType(BasicType value) {
-        if (value != null) {
-            switch (value) {
-            case BOOLEAN:
-                return Checkbox;
-            case CURRENCY:
-                return Currency;
-            case DATE:
-                return Date;
-            case DATE_TIME:
-                return DateTime;
-            case ID:
-                return Lookup;
-            case DECIMAL:
-            case DOUBLE:
-            case INTEGER:
-            case LONG:
-                return Number;
-            case STRING:
-                return Text;
-            case TIME:
-                return Time;
-            default:
-                break;
-            }
-        }
-        LOG.debug("Unable to determine DataType of {}", value);
-        return Unknown;
     }
 
     /**